+++ Web/[Features]/Inventory/Manage/Delete/Models.cs
@@ -1,6 +1,9 @@
-﻿namespace Inventory.Manage.Delete;
+﻿using System.ComponentModel;
+
+namespace Inventory.Manage.Delete;
 
 public class Request
 {
+    [DefaultValue("test default val")]
     public string ItemID { get; set; }
 }